% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation_weightHistory.R
\name{add_bias_f}
\alias{add_bias_f}
\title{Region of the cohort x age plane measurements (current, 1yr or 10yrs ago) pertain to}
\usage{
add_bias_f(d_a, by_sex_a = T, Glist_a)
}
\arguments{
\item{d_a}{data.table}

\item{by_sex_a}{whether the bias is sex-specific (\code{TRUE}) or not (\code{FALSE})}

\item{Glist_a}{list of models to use to compute bias}
}
\value{
0n for correct execution
}
\description{
Region of the cohort x age plane measurements (current, 1yr or 10yrs ago) pertain to
}
