% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{add_weights_f}
\alias{add_weights_f}
\title{Compute weights for the analysis of all nhanes cycles taken together}
\usage{
add_weights_f(data_a)
}
\arguments{
\item{data_a}{a \code{data.table} of nhanes data, including weights 
(either \code{demo_wtmec2yr} or \code{demo_wtmec4yr})}
}
\value{
None.
}
\description{
Compute weights for the analysis of all nhanes cycles taken together
}
\details{
Weights used are MEC weights, since the analysis involves measures
taken during MEC examination (e.g. measured height), and is therefore 
restricted to people both interviewed and MEC examined (RIDSTATR==2, which
has become examined == 1)
}
