% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_sigma_f.R
\name{analyze_sigma_f}
\alias{analyze_sigma_f}
\title{Plot variance-covariance matrix for each stratum}
\usage{
analyze_sigma_f(strata_a, RS_a)
}
\arguments{
\item{strata_a}{stratum individuals belong to .}

\item{result_set_a}{set of parameters analyzed.}
}
\value{
None.
}
\description{
Plot variance-covariance matrix for each stratum
}
