% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{clean_nhanes_f}
\alias{clean_nhanes_f}
\title{Clean nhanes data}
\usage{
clean_nhanes_f(data_a, height_a = "measured", verbose_a = F)
}
\arguments{
\item{data_a}{a \code{data.table} of nhanes data.}

\item{height_a}{which height to use in BMI computation. Either "measured" or "self-reported".}

\item{verbose_a}{whether details should be printed to screen.}
}
\value{
0 for correct execution
}
\description{
Clean nhanes data
}
\details{
This function : 
\itemize{
\item{changes some variable names;}
\item{changes values of sex : 1->m and 2-> f;}
\item{converts self-reported weights and height in kgs and cms;}
\item{adds new variable (current bmi, etc.);}
\item{group non-Mexican and Mexican Hispanics (new \code{race} variable);}
\item{group high-school graduate and less (newd \code{educ} variable)}
}
NB : no individual is removed from dataset.
}
