% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{clear_labels}
\alias{clear_labels}
\title{Remove "labelled" from class attribute}
\usage{
clear_labels(x)
}
\arguments{
\item{x}{data.table or list of data.tables}
}
\value{
A \code{data.table}
}
\description{
Remove "labelled" from class attribute
}
