% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{col_f}
\alias{col_f}
\title{Get colours for males and females}
\usage{
col_f(sex_a = "f", a_a = 50)
}
\arguments{
\item{sex_a}{Either  "f" or "m"}

\item{a_a}{alpha value}
}
\value{
rgb value
}
\description{
Get colours for males and females
}
