% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_Zage_f.R
\name{complete_Zage_f}
\alias{complete_Zage_f}
\title{Complete Zage series}
\usage{
complete_Zage_f(stratum_a, RS_a)
}
\arguments{
\item{stratum_a}{vector of strata for which imputation is potentially needed.}

\item{RS_a}{set of parameters analyzed.}
}
\value{
\code{0} for correct execution.
}
\description{
In case some ages were not attained and thus not reconstructed, then impute them.
}
