% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{core_res_name_f}
\alias{core_res_name_f}
\title{Returns the core character string of the files of results' names}
\usage{
core_res_name_f(RS_a, pt_a = "/")
}
\arguments{
\item{RS_a}{set of parameters analyzed.}

\item{pt_a}{separator.}
}
\value{
A character string.
}
\description{
Returns the core character string of the files of results' names
}
