% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_bfda.R
\name{export_bfda_f}
\alias{export_bfda_f}
\title{Prepare dataset used in functional data analysis}
\usage{
export_bfda_f(stratum_a, data_a, directory_a)
}
\arguments{
\item{stratum_a}{which stratum to save}

\item{data_a}{a \code{data.table}}

\item{directory_a}{where to save the \code{csv} file}
}
\value{
0 for correct execution
}
\description{
Prepare dataset used in functional data analysis
}
\details{
The dataset saved is restricted to columns that are absolutely needed by Matlab to avoid writing
}
