% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation_weightHistory.R
\name{G25_f}
\alias{G25_f}
\title{Function to estimate BMI at 25 based on measured data}
\usage{
G25_f(sex_a = c("m", "f"), w_a = "original", k_a = 6, nhanes25_a)
}
\arguments{
\item{sex_a}{c("m", "f"), "m" or "f"}

\item{w_a}{weight argument : "original" or "rescaled"}

\item{k_a}{dimension of the smooth term's basis}

\item{nhanes25_a}{dataset used}
}
\value{
A Generalized Additive Model
}
\description{
Function to estimate BMI at 25 based on measured data
}
