% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation_weightHistory.R
\name{G_f}
\alias{G_f}
\title{Estimation of the BMI(age, yob) surface using the different measures}
\usage{
G_f(
  type_recall_a = "curr",
  sex_a = c("m", "f"),
  weighted_a = T,
  K_a = 3,
  data_a,
  include_nhanesIII = T,
  dataIII_a = NULL
)
}
\arguments{
\item{sex_a}{c("m", "f"), "m" or "f"}

\item{weighted_a}{if true, use survey weights}

\item{K_a}{dimension of each basis of the tensor product smooth}

\item{data_a}{continuous NHANES data}

\item{include_nhanesIII}{If true, NHANES III data is also used in the estimation}

\item{dataIII_a}{NHANES III data}

\item{recall_a}{type of recall}
}
\value{
A Generalized Additive Model
}
\description{
Estimation of the age x yob -> E(BMI | age, yob) mapping using either current BMI, BMI 1y ago or BMI 10y ago
}
\details{
Pregnant women are removed if recall_a == 0 (current BMI),  
because past weights are asked before pregnancy
}
