% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_functions.R
\name{graphs_prev_f}
\alias{graphs_prev_f}
\title{Plot prevalence of obesity by cohort of birth}
\usage{
graphs_prev_f(
  prev_a,
  bmi_l_a,
  sex_a,
  TITLE_a = "",
  xlim_L,
  xlim_U,
  present_L = NULL,
  present_U = NULL
)
}
\arguments{
\item{prev_a}{matrix of a list returned by \code{\link{prevalence_f}}}

\item{bmi_l_a}{which bmi limit (should be present in \code{prev_a})}

\item{sex_a}{either "f" or "m"}

\item{TITLE_a}{title given to the plot}

\item{xlim_L}{lower limit of xaxis (reversed)}

\item{xlim_U}{upper limit of xaxis (reversed)}

\item{present_L}{lower limit of 'present' line}

\item{present_U}{upper limit of 'present' line}
}
\value{
ggplot object
}
\description{
Plot prevalence of obesity by cohort of birth
}
