% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{import_nhanes_f}
\alias{import_nhanes_f}
\title{Load NHANES data}
\usage{
import_nhanes_f(directory_a, path_a = "")
}
\arguments{
\item{directory_a}{directory of data}

\item{path}{path to \code{directory_a}}
}
\value{
A \code{data.table}
}
\description{
Load NHANES data
}
