% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation_weightHistory.R
\name{isObs_f}
\alias{isObs_f}
\title{Region of the cohort x age plane measurements (current, 1yr or 10yrs ago) pertain to}
\usage{
isObs_f(
  recall_a,
  minYr_a = 1999,
  maxYr_a = 2018,
  cohort_span_a = 1940:1995,
  age_span_a = 15:60
)
}
\arguments{
\item{recall_a}{type of recall}

\item{minYr_a}{first year of observation}

\item{maxYr_a}{last year of observation}

\item{cohort_span_a}{cohorts considered}

\item{age_span_a}{age span considered}
}
\value{
A matrix of booleans
}
\description{
Region of the cohort x age plane measurements (current, 1yr or 10yrs ago) pertain to
}
\details{
\code{recall_a} = 0 (current_bmi), 1 (bmi 1y ago) or 10 (bmi 10y ago)
}
