% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_results_f.R
\name{load_results_f}
\alias{load_results_f}
\title{Load results from BFDA}
\usage{
load_results_f(
  stratum_a,
  RS_a,
  res_type_a = "short",
  var_a,
  var_type_a = "csv"
)
}
\arguments{
\item{stratum_a}{vector of strata for which results are required}

\item{RS_a}{set of parameters analyzed}

\item{res_type_a}{either \code{long} (individual level posterior samples) 
or \code{short} (e.g. mean posterior values)}

\item{var_a}{the variable to be retrieved}

\item{var_type_a}{the file format (either \code{mat} or \code{csv})}
}
\value{
Either data.table of results or list returned by \code{R.matlab::readMat}
}
\description{
Load results from the folders storing the results of the functional data analysis
}
