% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_bfda.R
\name{make_bfda_data_f}
\alias{make_bfda_data_f}
\title{Prepare dataset used in functional data analysis}
\usage{
make_bfda_data_f(data_a, ages_study_a, cohorts_study_a)
}
\arguments{
\item{data_a}{list of \code{data.table}s}

\item{ages_study_a}{cohorts considered}

\item{cohorts_study_a}{age span considered}
}
\value{
A \code{data.table}
}
\description{
Prepare dataset used in functional data analysis
}
\details{
This function calls \code{\link{melt_nhanes_f}} and returns a data.table in long format where :
     \itemize{ 
     \item{only those aged ageMin to ageMax are kept,}
     \item{only those MEC examined are kept,}
     \item{those with missing educ, bmi and smoking are excluded.}
     }
}
