% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{make_char_pred_f}
\alias{make_char_pred_f}
\title{Return the name of the overweight/obesity category passed as an argument}
\usage{
make_char_pred_f(bmi_limit_a)
}
\arguments{
\item{bmi_limit_a}{BMI limit (25, 30, 35, 40).}
}
\value{
A character string.
}
\description{
Return the name of the overweight/obesity category passed as an argument
}
