% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{melt_nhanes_f}
\alias{melt_nhanes_f}
\title{NHANES data in long format (all cycles grouped)}
\usage{
melt_nhanes_f(input_data_a, verbose_a = T, bmis_a = "all")
}
\arguments{
\item{input_data_a}{NHANES data in wide format}

\item{verbose_a}{print details to screen}

\item{bmis_a}{wich BMI series are available and should be melted}
}
\value{
A \code{data.table}
}
\description{
NHANES data in long format (all cycles grouped)
}
