% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muSRES_all_f.R
\name{muSRES_all_f}
\alias{muSRES_all_f}
\title{Plot all average posterior mean BMI functions for given sex}
\usage{
muSRES_all_f(sex_a, RS_a, strata_a)
}
\arguments{
\item{sex_a}{Either female (\code{"f"}) or male (\code{"m"})}

\item{RS_a}{set of parameters analyzed}

\item{strata_a}{strata analyzed}
}
\value{
None
}
\description{
Should only be run if strata are defined based on :
Sex, Race, Education and Smoking
}
