% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muSRES_interactive_f.R
\name{muSRES_interactive_f}
\alias{muSRES_interactive_f}
\title{Interactive plot of average posterior mean BMI functions}
\usage{
muSRES_interactive_f(sex_a, RS_a, strata_a)
}
\arguments{
\item{sex_a}{Either female (\code{"f"}) or male (\code{"m"})}

\item{RS_a}{set of parameters analyzed}

\item{strata_a}{strata analyzed}
}
\value{
plotly visualization.
}
\description{
Should only be run if strata are defined based on :
Sex, Race, Education and Smoking
}
