% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muSRES_race_f.R
\name{muSRES_race_f}
\alias{muSRES_race_f}
\title{Plot average posterior mean BMI functions for all races, given sex,
education and smoking status}
\usage{
muSRES_race_f(sex_a, educ_a, smok_a, RS_a, strata_a, CI_a = F)
}
\arguments{
\item{sex_a}{\code{"f"} or \code{"m"}}

\item{educ_a}{educational attainment}

\item{smok_a}{\code{0} or \code{1}}

\item{RS_a}{set of parameters analyzed}

\item{strata_a}{strata analyzed}
}
\value{
None
}
\description{
Should only be run if strata are defined based on :
Sex, Race, Education and Smoking
}
