% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_functions.R
\name{obese_prop_f}
\alias{obese_prop_f}
\title{Compute proportion obese in each posterior sample}
\usage{
obese_prop_f(seqn_a, mcmc_a, limit_a, weighted_a = T)
}
\arguments{
\item{seqn_a}{ids of individuals for which proportion should be computed.}

\item{mcmc_a}{matrix of posterior samples (rows : individuals ; columns : samples).}

\item{limit_a}{lower limit of the obesity category investigated.}

\item{weighted_a}{whether proportions should be weighted using sample weighted.}
}
\value{
Vector of proportions in each posterior sample.
}
\description{
Compute proportion obese in each posterior sample
}
