% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{path_data_f}
\alias{path_data_f}
\title{Return the path to a csv file used by bfda}
\usage{
path_data_f(RS_a, stratum_a = NULL)
}
\arguments{
\item{RS_a}{set of parameters analyzed.}

\item{stratum_a}{vector of strata for which imputation is potentially needed. If \code{NULL}, all strata.}
}
\value{
A character string.
}
\description{
Return the path to a csv file used by bfda
}
