% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_traj_f.R
\name{plot_traj_f}
\alias{plot_traj_f}
\title{Plot trajectories}
\usage{
plot_traj_f(
  seqn_v_a,
  result_set_a,
  data_a,
  Z_a,
  Z_CL_a,
  Z_UL_a,
  bmi_a = "Deb2",
  ylim_a = c(15, 48),
  verbose_a = F,
  print_seqn_a = F
)
}
\arguments{
\item{seqn_v_a}{identifier (seqn) of individuals for which trajectories should be plotted}

\item{result_set_a}{set of parameters for which results are analyzed}

\item{data_a}{data.table that was used in BFDA}

\item{Z_a}{mean of posterior distribution}

\item{Z_CL_a}{lower limit of 95\%CI}

\item{Z_UL_a}{upper limit of 95\%CI}

\item{bmi_a}{\code{Uncorr}, \code{Deb1} or \code{Deb2}}

\item{ylim_a}{\code{ylim} for plot}

\item{verbose_a}{also print info to screen}

\item{print_seqn_a}{if \code{TRUE}, add \code{seqn} to plot}
}
\value{
None
}
\description{
Plot BMI trajectories of specific individuals
}
\examples{
plot_traj_f("seqn1111", RS, nhanes_bfda, Z, Z_CL, Z_UL)

}
