% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_functions.R
\name{prevalence_f}
\alias{prevalence_f}
\title{Compute age-specific prevalence of obesity}
\usage{
prevalence_f(PP_a, data_a, strata_a, RS_a, keyby_a = c("sex", "cob"))
}
\arguments{
\item{PP_a}{parameters for prediction, list of 3: \code{age_proj}, \code{age_span}, \code{bmi_limit}}

\item{data_a}{a \code{data.table} storing original data}

\item{strata_a}{strata analyzed}

\item{RS_a}{set of parameters analyzed}

\item{keyby_a}{factors defining groups for which prevalence is to be computed}
}
\value{
A list, including (first element) a list of matrices storing 
prevalences of obesity (one / value in \code{PP_a$bmi_limit})
}
\description{
Compute age-specific prevalence of obesity
}
