% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{seqn_subset_f}
\alias{seqn_subset_f}
\title{Create a vector of seqn according to cohort of birth}
\usage{
seqn_subset_f(beg_a, end_a, sex_a = NULL, race_a = NULL, educ_a = NULL, data_a)
}
\arguments{
\item{beg_a}{first year of birth included.}

\item{end_a}{last year of birth included.}

\item{sex_a}{restrict to this sex if given.}

\item{race_a}{restrict to this race/ethnicity if given.}

\item{educ_a}{restrict to this education group if given.}
}
\value{
Vector of seqn of individuals born \code{beg_a}-\code{end_a}.
}
\description{
Create a vector of seqn according to cohort of birth
}
