% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_functions.R
\name{store_prevalence_f}
\alias{store_prevalence_f}
\title{Get mean and quantiles of results by group}
\usage{
store_prevalence_f(x)
}
\arguments{
\item{x}{a matrix of posterior samples}
}
\value{
a data.table with mean, 2.5\% and 97.5\% quantiles of posterior distribution
}
\description{
Get mean and quantiles of results by group
}
