% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso_f.R
\name{tso_f}
\alias{tso_f}
\title{Compute weighted total time spent obese by members of a stratum}
\usage{
tso_f(stratum_a, data_a, result_set_a, limit_a, N_iterations_a = 10000L)
}
\arguments{
\item{stratum_a}{ids of individuals for which proportion should be computed.}

\item{data_a}{data.table of individuals studied (contains \code{seqn}, \code{weight}).}

\item{result_set_a}{set of parameters analyzed.}

\item{limit_a}{vector of lower limits of the obesity categories investigated.}
}
\value{
A matrix.
}
\description{
For each limit \code{l} in \code{limit_a}, the time spent above \code{l} (weighted using
survey weights) is computed for each individual of  \code{stratum_a} in each posterior sample. Results for each element of
\code{limit_a} are the stacked. The matrix returned thus has
\code{N_individuals x length(limit)} rows and \code{N_iterations_a} columns.
}
